# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: DocFlow API
  version: 1.0.0
  description: |
    DocFlow 文档管理服务 API。

    ## 认证
    - 业务端点要求 `Authorization: Bearer <access_token>`（HS256 JWT，subject 为用户 UUID）。
    - 登录/刷新通过 HttpOnly Cookie `refresh_token`（Path=/api/v1/auth/refresh，SameSite=Lax）轮换会话。

    ## 错误与限流
    - 错误响应统一为 `{error, code?}`；`code` 为机器可读错误码（如 NAME_CONFLICT、QUARANTINED）。
    - 认证接口按 user_id+IP 每分钟限流；登录按 IP+邮箱前缀哈希更严格限流；
      公开分享接口按 IP 独立限流；超限返回 429 并带 Retry-After。
    - 文件元数据响应带 ETag（基于 updated_at）。
servers:
  - url: /api/v1
    description: API 基础路径（契约测试据此归一化路由）
security:
  - bearerAuth: []
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refresh
      summary: 轮换刷新令牌
      description: >-
        从 refresh_token Cookie 读取刷新令牌并轮换（rotate）：旧令牌立即失效，
        新 Cookie 与新访问令牌随响应返回。该端点无独立限流。同源校验
        （CSRF，设计 6.1.5）：Origin/Referer 至少其一存在且 host 与请求 Host
        一致；严格模式（CSRF_STRICT，默认 true）下缺失两头一律 403——现代
        浏览器对 POST 均携带 Origin，同源 fetch 自然满足；非浏览器客户端
        （curl 等）需配置 CSRF_STRICT=false。账号被禁用 401（撤销刚轮换出的
        会话）；账号处于登录失败锁定期 423 ACCOUNT_LOCKED（同样撤销）。
      security: []
      parameters:
        - name: refresh_token
          in: cookie
          required: true
          schema: { type: string }
          description: HttpOnly 刷新令牌 Cookie
      responses:
        '200':
          description: 新访问令牌；Set-Cookie 轮换 refresh_token
          headers:
            Set-Cookie:
              description: 轮换后的 refresh_token Cookie
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          description: 刷新令牌缺失、无效或已过期；或账号已禁用
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: >-
            同源校验失败（CSRF）：Origin/Referer 缺失（严格模式）或 host 与
            请求 Host 不一致，body 为 {error, code: CSRF_REJECTED}。
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '423':
          description: '账号处于登录失败锁定期，{error, code: ACCOUNT_LOCKED}'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      operationId: forgotPassword
      summary: 请求密码重置邮件
      description: >-
        邮箱存在活跃账号时创建 30 分钟有效的一次性重置令牌并发送邮件
        （链接 {PUBLIC_BASE_URL}/reset/{token}）；无论邮箱是否存在一律 202，
        不泄露账号存在性。按 IP 独立严格限流（10/min）。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
      responses:
        '202':
          description: 已受理（不区分邮箱是否存在）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: accepted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: 凭一次性令牌重置密码
      description: >-
        消费一次性重置令牌（原子标记 used_at），更新密码并撤销该用户全部
        会话。令牌未知/已使用/已过期 400。按 IP 独立严格限流（10/min）。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '204':
          description: 已重置（全部会话失效，需重新登录）
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/change-password:
    post:
      tags: [Auth]
      operationId: changePassword
      summary: 修改密码（认证）
      description: >-
        校验旧密码（错误 403）与新密码强度后更新哈希，撤销该用户全部其他
        会话，并轮换当前会话——响应头 Set-Cookie 下发本会话新的 refresh_token。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '204':
          description: 已修改；Set-Cookie 轮换本会话 refresh_token
          headers:
            Set-Cookie:
              description: 轮换后的 refresh_token Cookie（其他会话已被撤销）
              schema: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: 旧密码错误
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /folders/{id}/acl:
    get:
      tags: [Spaces]
      operationId: getFolderACL
      summary: 查看目录的路径级 ACL 条目
      description: >-
        返回空间作用域目录（含空间根目录）的全部 folder_acl 条目（附 subject
        名称解析：user → username、space → 空间名）。仅文件夹所在空间
        owner/admin 或系统 admin 可读；非空间作用域目录 400（NOT_SPACE_FOLDER）。
        求值语义（read/write/delete/share 判定先走 ACL 链）见 PUT 端点描述。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 目录 ID（空间根目录或其子目录）
      responses:
        '200':
          description: ACL 条目列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FolderACL' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    put:
      tags: [Spaces]
      operationId: replaceFolderACL
      summary: 整体替换目录的 ACL 条目（空数组即清空）
      description: >-
        整体替换目录的全部 folder_acl 条目（事务内删旧插新）。仅空间作用域
        目录可设置（非空间目录 400）；权限：文件夹所在空间 owner/admin 或
        系统 admin。求值规则：沿 parent 链自目标文件/目录向上到空间根收集
        条目，最近节点优先；同节点 user 条目优先于 space 条目、deny 优先于
        allow；链上无适用条目（含未设置任何 ACL 的目录）回退空间角色判定
        （直接成员 ∪ 用户组取最高）——无 ACL 时行为与既有权限完全一致。
        审计 acl.update。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 目录 ID（空间根目录或其子目录）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FolderACLUpdate' }
      responses:
        '200':
          description: 替换后的条目列表（与 GET 同构）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FolderACL' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Authorization: Bearer <token>，两类凭证等价可用：
        ① HS256 JWT（subject 为用户 UUID，过期时间以服务配置为准）；
        ② PAT 个人访问令牌：dfpat_ + 43 字符 URL-safe base64（经
        POST /tokens 创建，撤销/过期即失效；无 refresh 语义）。
  responses:
    BadRequest:
      description: 请求参数或请求体非法
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: 缺失或无效的访问令牌
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: 无权限执行该操作（含只读目录、根目录保护、对象不可用等）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: 资源不存在或不属于当前用户（不泄露存在性）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: 触发限流；响应头 Retry-After 为当前窗口剩余秒数
      headers:
        Retry-After:
          description: 建议等待秒数
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 人类可读错误信息
        code:
          type: string
          description: >-
            机器可读错误码；当前实现使用 NAME_CONFLICT / NOT_DELETED / PARENT_DELETED /
            QUARANTINED / FAILED / TOTP_REQUIRED（POST /auth/login 对已启用两步
            验证账号的 401，error=totp_required）/ ACCOUNT_LOCKED（登录与
            refresh 遇锁定期账号的 423）/ QUOTA_EXCEEDED（上传建会话配额
            超限的 403）/ CSRF_REJECTED（refresh/logout 同源校验失败的 403）
        status:
          type: string
          description: 上传相关错误附带的会话状态（仅 uploads complete 端点）
...
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          description: 一次性重置令牌（重置邮件 /reset/{token} 链接中的明文，30 分钟有效）
        password:
          type: string
          format: password
          description: 至少 12 字符，须含大写、小写与数字
    ChangePasswordRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password: { type: string, format: password }
        new_password:
          type: string
          format: password
          description: 至少 12 字符，须含大写、小写与数字
...
    TokenPair:
      type: object
      required: [access_token, token_type]
      properties:
        access_token: { type: string, description: JWT 访问令牌 }
        token_type: { type: string, const: Bearer }
...
    FolderACLEntry:
      type: object
      description: >-
        folder_acl 条目（migration 031，设计 6.5.3/6.5.4 最小落地）：
        (目录, subject, effect, permissions) 一行一条；UNIQUE(folder_id,
        subject_type, subject_id)。subject_name 为服务端解析的展示名
        （user → username、space → 空间名），解析失败为空串。
      properties:
        subject_type: { type: string, enum: [user, space] }
        subject_id: { type: string, format: uuid }
        subject_name: { type: string, description: 展示名（username / 空间名） }
        effect: { type: string, enum: [allow, deny] }
        permissions:
          type: array
          minItems: 1
          items: { type: string, enum: [read, write, delete, share] }
        created_at: { type: string, format: date-time }
    FolderACL:
      type: object
      properties:
        folder_id: { type: string, format: uuid }
        entries:
          type: array
          items: { $ref: '#/components/schemas/FolderACLEntry' }
    FolderACLUpdate:
      type: object
      description: PUT /folders/{id}/acl 请求体：整体替换（空数组即清空）
      required: [entries]
      properties:
        entries:
          type: array
          items:
            type: object
            required: [subject_type, subject_id, effect, permissions]
            properties:
              subject_type: { type: string, enum: [user, space] }
              subject_id: { type: string, format: uuid }
              effect: { type: string, enum: [allow, deny] }
              permissions:
                type: array
                minItems: 1
                items: { type: string, enum: [read, write, delete, share] }